requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
wordcloud>=1.9.2
matplotlib>=3.7.0